==> South Sudan resilience/scripts/05 South Sudan resilience panel survey - master.R <==
# South Sudan panel survey
# master

source("scripts/0 South Sudan resilience panel survey - prep.r")

source("scripts/02 South Sudan resilience panel survey - data prep.r")

source("scripts/03 South Sudan resilience panel survey - indicators.r")

#rmarkdown::render("scripts/03 South Sudan resilience panel survey - indicators.rmd")
